extracted EquationSet class from equation.py into a new file
